+++ src/Smartstore.Core/Migrations/SmartDbContextDataSeeder.cs
@@ -1,4 +1,8 @@
-﻿using Smartstore.Data.Migrations;
+﻿using Newtonsoft.Json.Linq;
+using NuGet.ContentModel;
+using System.Runtime.ConstrainedExecution;
+using System.Xml.Linq;
+using Smartstore.Data.Migrations;
 
 namespace Smartstore.Core.Data.Migrations
 {
@@ -46,6 +50,42 @@ namespace Smartstore.Core.Data.Migrations
             builder.AddOrUpdate("Admin.Customers.Customers.Fields.StreetAddress2.Hint", "The address 2.", "Adresszusatz");
             builder.AddOrUpdate("Admin.Orders.Address.Address2", "Address 2", "Adresszusatz");
             builder.AddOrUpdate("PDFPackagingSlip.Address2", "Address 2: {0}", "Adresszusatz: {0}");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.SocialSettings.FlickrLink",
+                "Flickr Link",
+                "Flickr Link",
+                "Leave this field empty if the Flickr link should not be shown.",
+                "Lassen Sie dieses Feld leer, wenn der Flickr Link nicht angezeigt werden soll.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.SocialSettings.LinkedInLink",
+                "LinkedIn Link",
+                "LinkedIn Link",
+                "Leave this field empty if the LinkedIn link should not be shown.",
+                "Lassen Sie dieses Feld leer, wenn der LinkedIn Link nicht angezeigt werden soll.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.SocialSettings.XingLink",
+                "Xing Link",
+                "Xing Link",
+                "Leave this field empty if the Xing link should not be shown.",
+                "Lassen Sie dieses Feld leer, wenn der Xing Link nicht angezeigt werden soll.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.SocialSettings.TikTokLink",
+                "TikTok Link",
+                "TikTok Link",
+                "Leave this field empty if the TikTok link should not be shown.",
+                "Lassen Sie dieses Feld leer, wenn der TikTok Link nicht angezeigt werden soll.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.SocialSettings.SnapchatLink",
+                "Snapchat Link",
+                "Snapchat Link",
+                "Leave this field empty if the Snapchat link should not be shown.",
+                "Lassen Sie dieses Feld leer, wenn der Snapchat Link nicht angezeigt werden soll.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.SocialSettings.VimeoLink",
+                "Vimeo Link",
+                "Vimeo Link",
+                "Leave this field empty if the Vimeo link should not be shown.",
+                "Lassen Sie dieses Feld leer, wenn der Vimeo Link nicht angezeigt werden soll.");
         }
     }
 }